Move all "leaf" modules into lib/
