+++ packages/liveblocks-core/src/lib/EventSource.ts
