+++ packages/liveblocks-core/src/lib/__tests__/shallow.test.ts
