+++ packages/liveblocks-core/src/lib/deprecation.ts
@@ -1,4 +1,4 @@
-import * as console from "./lib/fancy-console";
+import * as console from "./fancy-console";
 
 /**
  * Tools to help with the controlled deprecation of public APIs.
