+++ packages/liveblocks-core/src/room.ts
@@ -2,9 +2,9 @@ import type { ApplyResult, ManagedPool } from "./AbstractCrdt";
 import { OpSource } from "./AbstractCrdt";
 import type { RoomAuthToken } from "./AuthToken";
 import { isTokenExpired, parseRoomAuthToken, RoomScope } from "./AuthToken";
-import type { Callback, Observable } from "./EventSource";
-import { makeEventSource } from "./EventSource";
 import { assertNever, nn } from "./lib/assert";
+import type { Callback, Observable } from "./lib/EventSource";
+import { makeEventSource } from "./lib/EventSource";
 import * as console from "./lib/fancy-console";
 import { LiveObject } from "./LiveObject";
 import { MeRef } from "./MeRef";
